перенос добавление лаб. ссылки в отдельынй компонент
